pull out interface to start to create public interface
